Add the Scroll Concierge™. (#266)
* Initial version of the Scroll Concierge™.

* Allow debugger for use in development.

* Use display-table to prevent "shifting" in Lede Banner.

Flexbox has some weirdness with how layout renders (it'll first build the one column, and then reflow once it "hits" the 2nd column: https://goo.gl/SpXOyR). This was causing the offset of the "concierge" to shift by the time we scrolled there.

* Add some fallbacks for League Gothic.

This will help minimize the "flash" of unstyled text that you otherwise see while the webfont is loading.

* Update scroll concierge to use RAF.

* Wait for headline font to be loaded before scrolling.

The fallback font for this often has a fairly different height than League Gothic so the layout shifts. We want to wait for that "final" layout before caclulating the scroll offset.

* Allow the user to interrupt the scroll.